HungryMe_Backend_Django/requirements.txt
boto3==1.14.3
botocore==1.17.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-cors-headers==3.2.1
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
django-heroku==0.3.1
docutils==0.15.2
gunicorn==19.9.0
graphene-django==2.11.0
idna==2.9
jmespath==0.10.0
Pillow==7.1.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.8
whitenoise==2.0.6
